Interface and it's implementations.
